+++ app/src/main/java/ru/geekbrains/myweatherapp/forecastdata/onecallapi/Hourly.java
@@ -42,7 +42,7 @@ public class Hourly {
     private List<HourlyWeather> weather = null;
     @SerializedName("pop")
     @Expose
-    private Integer pop;
+    private Double pop;
 
     public Integer getDt() {
         return dt;
@@ -132,11 +132,11 @@ public class Hourly {
         this.weather = weather;
     }
 
-    public Integer getPop() {
+    public Double getPop() {
         return pop;
     }
 
-    public void setPop(Integer pop) {
+    public void setPop(double pop) {
         this.pop = pop;
     }
